refactor: teamId 사용하는 API 호출함수의 teamId 전역 상태로 리팩토링 [TM-144]
