+++ src/features/resource/model/useResourceQueries.ts
@@ -1,17 +1,19 @@
-import apiRequest from '@/shared/api/apiRequest';
-import { TEAM_ID } from '@/shared/config/constants/team.constants';
-import { APIResponse } from '@/shared/types/api.types';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import apiRequest from '@/shared/api/apiRequest';
+import { useTeamStore } from '@/shared/model/store/teamStore';
+import { APIResponse } from '@/shared/types/api.types';
 
 /* 자료 업로드 */
 export const useUploadResource = () => {
+  const teamId = useTeamStore((state) => state.teamId);
+
   return useMutation<APIResponse<{ dataId: number }>, Error, File>({
     mutationFn: async (file: File) => {
       const formData = new FormData();
       formData.append('file', file);
 
       return await apiRequest({
-        url: `/api/v2/data/${TEAM_ID}`,
+        url: `/api/v2/data/${teamId}`,
         method: 'POST',
         data: formData,
         headers: {
